wheat_disease_system/wheat_disease/requirements.txt
# Core (required)
flask>=2.3.0
werkzeug>=2.3.0
pillow>=9.0.0
numpy>=1.21.0
reportlab>=3.6.0

# ML (required for real inference — install these!)
torch>=1.13.0
torchvision>=0.14.0
timm>=0.9.0

# Optional
# gunicorn>=20.1.0     # Production server
